OpenGL*: Rename ShadersGL*.cs to QuadShaderGeneratorGL*.cs
