+++ Furball.Vixie/Graphics/Backends/OpenGL41/QuadShaderGeneratorGL41.cs
@@ -2,10 +2,10 @@ using System.Collections.Immutable;
 using System.Text;
 using Furball.Vixie.Helpers;
 
-namespace Furball.Vixie.Graphics.Backends.OpenGL20 {
-    public static class ShadersGL20 {
+namespace Furball.Vixie.Graphics.Backends.OpenGL41 {
+    public static class QuadShaderGeneratorGL41 {
         public static string GetFragment(GraphicsBackend backend) {
-            string orig = ResourceHelpers.GetStringResource("ShaderCode/OpenGL20/FragmentShader.glsl");
+            string orig = ResourceHelpers.GetStringResource("ShaderCode/OpenGL41/InstancedRenderer/FragmentShader.glsl");
 
             StringBuilder uniformBuilder = new();
             StringBuilder ifBuilder = new();
@@ -15,7 +15,7 @@ namespace Furball.Vixie.Graphics.Backends.OpenGL20 {
 
                 if (i != 0) ifBuilder.Append("else ");
 
-                ifBuilder.Append($"if(tex_id == {i}) {{ gl_FragColor = texture2D(tex_{i}, fs_in_tex) * fs_in_col; }}");
+                ifBuilder.Append($"if(fs_in_texid == {i}) {{ OutputColor = texture(tex_{i}, fs_in_tex) * fs_in_col; }}");
             }
 
             return orig
